+++ native/core/src/execution/jni_api.rs
@@ -66,8 +66,8 @@ use crate::execution::memory_pools::{
 use crate::execution::operators::ScanExec;
 use crate::execution::shuffle::{read_ipc_compressed, CompressionCodec};
 use crate::execution::spark_plan::SparkPlan;
-#[cfg(feature = "tracing")]
-use crate::execution::utils::Recorder;
+
+use crate::execution::tracing::{trace_begin, trace_end};
 use log::info;
 use once_cell::sync::Lazy;
 
@@ -84,9 +84,6 @@ static TOKIO_RUNTIME: Lazy<Runtime> = Lazy::new(|| {
         .build()
         .expect("Failed to create Tokio runtime")
 });
-
-#[cfg(feature = "tracing")]
-pub(crate) static RECORDER: Lazy<Recorder> = Lazy::new(|| Recorder::new());
 
 fn parse_usize_env_var(name: &str) -> Option<usize> {
     std::env::var_os(name).and_then(|n| n.to_str().and_then(|s| s.parse::<usize>().ok()))
@@ -158,8 +155,7 @@ pub unsafe extern "system" fn Java_org_apache_comet_Native_createPlan(
     explain_native: jboolean,
 ) -> jlong {
     try_unwrap_or_throw(&e, |mut env| {
-        #[cfg(feature = "tracing")]
-        RECORDER.begin_task("create_plan");
+        trace_begin("create_plan");
 
         // Init JVM classes
         JVMClasses::init(&mut env);
@@ -241,8 +237,7 @@ pub unsafe extern "system" fn Java_org_apache_comet_Native_createPlan(
             memory_pool_config,
         });
 
-        #[cfg(feature = "tracing")]
-        RECORDER.end_task("create_plan");
+        trace_end("create_plan");
 
         Ok(Box::into_raw(exec_context) as i64)
     })
@@ -367,8 +362,7 @@ pub unsafe extern "system" fn Java_org_apache_comet_Native_executePlan(
     schema_addrs: jlongArray,
 ) -> jlong {
     try_unwrap_or_throw(&e, |mut env| {
-        #[cfg(feature = "tracing")]
-        RECORDER.begin_task("execute_plan");
+        trace_begin("execute_plan");
 
         // Retrieve the query
         let exec_context = get_execution_context(exec_context);
@@ -430,8 +424,7 @@ pub unsafe extern "system" fn Java_org_apache_comet_Native_executePlan(
                 Poll::Ready(Some(output)) => {
                     // prepare output for FFI transfer
 
-                    #[cfg(feature = "tracing")]
-                    RECORDER.end_task("execute_plan");
+                    trace_end("execute_plan");
 
                     return prepare_output(
                         &mut env,
@@ -457,8 +450,7 @@ pub unsafe extern "system" fn Java_org_apache_comet_Native_executePlan(
                         }
                     }
 
-                    #[cfg(feature = "tracing")]
-                    RECORDER.end_task("execute_plan");
+                    trace_end("execute_plan");
 
                     return Ok(-1);
                 }
@@ -561,8 +553,7 @@ pub unsafe extern "system" fn Java_org_apache_comet_Native_writeSortedFileNative
     enable_fast_encoding: jboolean,
 ) -> jlongArray {
     try_unwrap_or_throw(&e, |mut env| unsafe {
-        #[cfg(feature = "tracing")]
-        RECORDER.begin_task("writeSortedFileNative");
+        trace_begin("writeSortedFileNative");
 
         let data_types = convert_datatype_arrays(&mut env, serialized_datatypes)?;
 
@@ -627,8 +618,7 @@ pub unsafe extern "system" fn Java_org_apache_comet_Native_writeSortedFileNative
         let long_array = env.new_long_array(2)?;
         env.set_long_array_region(&long_array, 0, &[written_bytes, checksum])?;
 
-        #[cfg(feature = "tracing")]
-        RECORDER.end_task("writeSortedFileNative");
+        trace_end("writeSortedFileNative");
 
         Ok(long_array.into_raw())
     })
@@ -664,16 +654,14 @@ pub unsafe extern "system" fn Java_org_apache_comet_Native_decodeShuffleBlock(
     schema_addrs: jlongArray,
 ) -> jlong {
     try_unwrap_or_throw(&e, |mut env| {
-        #[cfg(feature = "tracing")]
-        RECORDER.begin_task("decodeShuffleBlock");
+        trace_begin("decodeShuffleBlock");
 
         let raw_pointer = env.get_direct_buffer_address(&byte_buffer)?;
         let length = length as usize;
         let slice: &[u8] = unsafe { std::slice::from_raw_parts(raw_pointer, length) };
         let batch = read_ipc_compressed(slice)?;
 
-        #[cfg(feature = "tracing")]
-        RECORDER.end_task("decodeShuffleBlock");
+        trace_end("decodeShuffleBlock");
 
         prepare_output(&mut env, array_addrs, schema_addrs, batch, false)
     })
