+++ native/core/src/execution/utils.rs
@@ -23,8 +23,6 @@ use arrow::{
     error::ArrowError,
     ffi::{from_ffi, FFI_ArrowArray, FFI_ArrowSchema},
 };
-#[cfg(feature = "tracing")]
-use datafusion::common::instant::Instant;
 
 impl From<ArrowError> for ExecutionError {
     fn from(error: ArrowError) -> ExecutionError {
@@ -128,46 +126,4 @@ pub fn bytes_to_i128(slice: &[u8]) -> i128 {
     }
 
     i128::from_le_bytes(bytes)
-}
-
-/// Log events using Chrome trace format JSON
-/// https://github.com/catapult-project/catapult/blob/main/tracing/README.md
-#[cfg(feature = "tracing")]
-pub struct Recorder {
-    now: Instant,
-}
-
-#[cfg(feature = "tracing")]
-impl Recorder {
-    pub fn new() -> Self {
-        print!("[ ");
-        Self {
-            now: Instant::now(),
-        }
-    }
-    pub fn begin_task(&self, name: &str) {
-        self.log_event(name, "B")
-    }
-
-    pub fn end_task(&self, name: &str) {
-        self.log_event(name, "E")
-    }
-
-    fn log_event(&self, name: &str, ph: &str) {
-        let thread_id = std::thread::current().id();
-
-        // TODO could be more efficient
-        let id_num: u64 = format!("{:?}", thread_id)
-            .trim_start_matches("ThreadId(")
-            .trim_end_matches(")")
-            .parse()
-            .unwrap();
-
-        println!(
-            "{{ \"name\": \"{}\", \"cat\": \"PERF\", \"ph\": \"{ph}\", \"pid\": 1, \"tid\": {}, \"ts\": {} }},",
-            name,
-            id_num,
-            self.now.elapsed().as_nanos()
-        );
-    }
 }
